refactor: streamline imports by modifying exports
