+++ src/pages/CreatePost/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './CreatePost';
