+++ src/pages/Introduction/index.ts
@@ -1,1 +1,1 @@
-export {}
+export * from './Introduction';
