+++ src/pages/index.ts
@@ -1,1 +1,3 @@
-export {}
+export * from './TestPage/TestPage';
+export * from './CreatePost/CreatePost';
+export * from './Introduction/Introduction';
