+++ src/router/Router.tsx
@@ -1,6 +1,5 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import { CreatePost } from '../pages/CreatePost/CreatePost';
-import { Introduction } from '../pages/Introduction/Introduction';
+import { CreatePost, Introduction, TestPage } from '../pages';
 
 export const RouterProvider = () => {
   return (
@@ -8,6 +7,7 @@ export const RouterProvider = () => {
       <Routes>
         <Route path="/" element={<Introduction />} />
         <Route path="/create" element={<CreatePost />} />
+        <Route path="/test" element={<TestPage />} />
       </Routes>
     </BrowserRouter>
   );
